+++ R/squishList.R
@@ -6,10 +6,11 @@
 #'
 #' @details items with the same name are assumed to have the same structure
 #'   and will be combined. Dataframes will be combined with bind_rows, vectors
-#'   just be collapsed into one vector, matrices will be combined with rbind, 
+#'   just be collapsed into one vector, matrices will be combined with rbind,
 #'   lists will be combined recursively with another call to \code{squishList}
 #'
 #' @param myList a list with named elements to be compressed
+#' @param unique logical flag to try and reduce result to only unique values
 #'
 #' @return a list with one element for every unique name in the original list
 #'
@@ -26,10 +27,10 @@
 #' myList <- list(a=list(c=1:2, d=2), b=letters[1:3], a=list(c=4:5, d=6:9))
 #' squishList(myList)
 #'
-#' @importFrom dplyr bind_rows
+#' @importFrom dplyr bind_rows distinct
 #' @export
 #'
-squishList <- function(myList) {
+squishList <- function(myList, unique=FALSE) {
     myNames <- unique(names(myList))
     if(is.null(myNames)) return(myList)
     result <- vector('list', length=length(myNames))
@@ -45,10 +46,14 @@ squishList <- function(myList) {
         } else if(all(sapply(thisNameData, function(x) inherits(x, 'list')))) {
             thisNameData <- unlist(thisNameData, recursive = FALSE)
             names(thisNameData) <- gsub(paste0(n, '\\.'), '', names(thisNameData))
-            squishList(thisNameData)
+            squishList(thisNameData, unique)
             # } else if(all(thisClasses=='data.frame')) {
         } else if(all(sapply(thisNameData, function(x) inherits(x, 'data.frame')))) {
-            bind_rows(thisNameData)
+            if(isTRUE(unique)) {
+                distinct(bind_rows(thisNameData))
+            } else {
+                bind_rows(thisNameData)
+            }
             # } else if(all(thisClasses=='NULL')) {
         } else if(all(sapply(thisNameData, function(x) inherits(x, 'NULL')))) {
             next
@@ -57,7 +62,11 @@ squishList <- function(myList) {
             do.call(rbind, thisNameData)
         } else {
             # thisNameData[[1]]
-            unlist(thisNameData, use.names = FALSE)
+            if(isTRUE(unique)) {
+                unique(unlist(thisNameData, use.names = FALSE))
+            } else {
+                unlist(thisNameData, use.names = FALSE)
+            }
         }
     }
     result