+++ tests/testthat/test-squish.R
@@ -21,10 +21,22 @@ test_that('Proper output from squishList', {
                      a=3:4)
     nullSquish <- squishList(nullList)
     expect_identical(nullSquish$a, 1:4)
-    
+
     matList <- list(a=matrix(1, nrow=1, ncol=2),
                     a=matrix(1, nrow=2, ncol=2))
     matSquish <- squishList(matList)
     expect_identical(matSquish$a, matrix(1, nrow=3, ncol=2))
     expect_identical(squishList(list()), list())
+    vecList <- list(a=1:4,
+                    b=data.frame(x=1:3),
+                    c=c('a', 'b'),
+                    a=2:5,
+                    b=data.frame(x=3:6),
+                    c=c('b', 'c'))
+    unqSquish <- squishList(vecList, unique=TRUE)
+    expect_identical(unqSquish$a, 1:5)
+    expect_identical(unqSquish$b, data.frame(x=1:6))
+    expect_identical(unqSquish$c, c('a', 'b', 'c'))
+    unqList <- list(a=1:10, b=data.frame(x=1:10), c=letters[1:5])
+    expect_identical(squishList(c(unqList, unqList), unique=TRUE), unqList)
 })